Renamed breaking-change-detector -> diff-processor (#8786)
* Renamed breaking-change-detector -> diff-processor

* Removed printf statements

* Removed debugging lines
