+++ .ci/scripts/go-plus/github-differ/generate_comment.sh
@@ -59,11 +59,11 @@ popd
 
 ## Breaking change setup and execution
 set +e
-pushd $MM_LOCAL_PATH/tools/breaking-change-detector
+pushd $MM_LOCAL_PATH/tools/diff-processor
 cp -r $TPG_LOCAL_PATH old/
 cp -r $TPG_LOCAL_PATH new/
 make build OLD_REF=$OLD_BRANCH NEW_REF=$NEW_BRANCH
-TPG_BREAKING="$(bin/breaking-change-detector)"
+TPG_BREAKING="$(bin/diff-processor)"
 retVal=$?
 if [ $retVal -ne 0 ]; then
     TPG_BREAKING=""
@@ -74,7 +74,7 @@ rm -rf ./old/ ./new/ ./bin/
 cp -r $TPGB_LOCAL_PATH old/
 cp -r $TPGB_LOCAL_PATH new/
 make build OLD_REF=$OLD_BRANCH NEW_REF=$NEW_BRANCH
-TPGB_BREAKING="$(bin/breaking-change-detector)"
+TPGB_BREAKING="$(bin/diff-processor)"
 retVal=$?
 if [ $retVal -ne 0 ]; then
     TPGB_BREAKING=""
