+++ tools/diff-processor/GNUmakefile
