+++ tools/diff-processor/README.md
@@ -1,0 +1,21 @@
+# Diff processor
+
+Provides tools that operate on diffs between provider versions.
+
+## Run
+
+```bash
+# set up old / new dirs
+make clone OWNER_REPO=modular-magician/terraform-provider-google
+
+# build based on old / new dirs
+make build OLD_REF=branch_or_commit NEW_REF=branch_or_commit
+
+# Run the binary
+bin/diff-processor
+```
+
+## Test
+```bash
+go test ./...
+```
