+++ tools/diff-processor/comparison.go
@@ -5,7 +5,7 @@ import (
 	oldProvider "google/provider/old/google"
 	"strings"
 
-	"github.com/GoogleCloudPlatform/magic-modules/.ci/breaking-change-detector/rules"
+	"github.com/GoogleCloudPlatform/magic-modules/.ci/diff-processor/rules"
 	"github.com/golang/glog"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
