+++ tools/diff-processor/constants/constants.go
