+++ tools/diff-processor/main.go
