+++ tools/diff-processor/rules/message_context.go
