+++ tools/diff-processor/rules/rule.go
