+++ tools/diff-processor/rules/rule_test.go
