+++ tools/diff-processor/rules/rules_provider.go
