+++ tools/diff-processor/rules/rules_resource_inventory.go
