+++ tools/diff-processor/rules/rules_resource_inventory_test.go
