+++ tools/diff-processor/rules/rules_resource_schema.go
