+++ tools/diff-processor/rules/rules_resource_schema_test.go
